ISIXE-1529 fixed some Windows warnings

+++ libs/i6engine-modules/include/i6engine/api/components/TerrainAppearanceComponent.h
@@ -123,7 +123,7 @@ namespace api {
 
 		void sendUpdateMessage();
 
-		void convertPositionToVertex(const Vec3 & pos, uint64_t & x, uint64_t & z) const;
+		void convertPositionToVertex(const Vec3 & pos, size_t & x, size_t & z) const;
 	};
 
 } /* namespace api */
