+++ libs/i6engine-modules/src/api/components/TerrainAppearanceComponent.cpp
@@ -112,7 +112,7 @@ namespace api {
 	}
 
 	void TerrainAppearanceComponent::setHeightAtPosition(const Vec3 & pos, double height) {
-		uint64_t x, z;
+		size_t x, z;
 		convertPositionToVertex(pos, x, z);
 		setHeightAtPosition(x, z, height);
 	}
@@ -122,7 +122,7 @@ namespace api {
 	}
 
 	double TerrainAppearanceComponent::getHeightAtPosition(const Vec3 & pos) const {
-		uint64_t x, z;
+		size_t x, z;
 		convertPositionToVertex(pos, x, z);
 		return getHeightAtPosition(x, z);
 	}
@@ -196,10 +196,10 @@ namespace api {
 		return result;
 	}
 
-	void TerrainAppearanceComponent::convertPositionToVertex(const Vec3 & pos, uint64_t & x, uint64_t & z) const {
+	void TerrainAppearanceComponent::convertPositionToVertex(const Vec3 & pos, size_t & x, size_t & z) const {
 		double edgeLength = _size / _vertices;
-		x = uint64_t((pos.getX() + _size * 0.5 + edgeLength * 0.5) / edgeLength - 0.5);
-		z = _vertices - uint64_t((pos.getZ() + _size * 0.5 + edgeLength * 0.5) / edgeLength + 0.5);
+		x = size_t((pos.getX() + _size * 0.5 + edgeLength * 0.5) / edgeLength - 0.5);
+		z = _vertices - size_t((pos.getZ() + _size * 0.5 + edgeLength * 0.5) / edgeLength + 0.5);
 	}
 
 } /* namespace api */
